Moved common test types to a seperate class to avoid code duplication
